# openapi.yaml
openapi: 3.0.3
info:
  title: University API
  version: "1.0.0"
  description: >
    API for user authentication, course management, grade management, and user management.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
paths:
...
  /courses/:
    get:
      tags:
        - Courses
      summary: List courses
      description: >
        **Requires JWT authentication.** Retrieve a paginated list of courses.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: The current page number.
        - in: query
          name: per_page
          schema:
            type: integer
            default: 10
          description: The number of courses per page.
      responses:
        "200":
          description: Courses retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Course'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      per_page:
                        type: integer
                      total_pages:
                        type: integer
                      total_items:
                        type: integer
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Courses
      summary: Create a new course
      description: >
        **Requires JWT authentication.** Only Administrators or Professors can create a course.
      security:
        - bearerAuth: []
      requestBody:
        description: Course creation payload.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Introduction to Programming
      responses:
        "200":
          description: Course created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Course created successfully.
                  course:
                    $ref: '#/components/schemas/Course'
        "400":
          description: Missing course name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /courses/{course_id}:
    get:
      tags:
        - Courses
      summary: Get course details
      description: >
        **Requires JWT authentication.** Retrieve details of a specific course.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_id
          required: true
          schema:
            type: string
          description: The unique identifier of the course.
      responses:
        "200":
          description: Course details returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  course:
                    $ref: '#/components/schemas/Course'
        "404":
          description: Course not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Courses
      summary: Update a course
      description: >
        **Requires JWT authentication.** Only Administrators or the assigned Professor may update the course.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_id
          required: true
          schema:
            type: string
          description: The unique identifier of the course.
      requestBody:
        description: JSON payload with the new course name.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Advanced Programming
      responses:
        "200":
          description: Course updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Course updated successfully.
                  course:
                    $ref: '#/components/schemas/Course'
        "400":
          description: Missing course name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Course not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Courses
      summary: Delete a course
      description: >
        **Requires JWT authentication.** Only Administrators can delete courses.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: course_id
          required: true
          schema:
            type: string
          description: The unique identifier of the course.
      responses:
        "200":
          description: Course deleted successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: Course deleted successfully.
        "403":
          description: Unauthorized access.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Course not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: string
          example: "course123"
        name:
          type: string
          example: Introduction to Programming
        professor_id:
          type: string
          example: "prof123"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
      required:
        - error